scripts/requirements.txt
jinja2
openshift
kubernetes
ansible
PyYAML==6.0.2
